+++ plugin-locationcomponent/src/main/java/com/mapbox/maps/plugin/locationcomponent/LocationComponentPluginImpl.kt
@@ -340,7 +340,7 @@ class LocationComponentPluginImpl : LocationComponentPlugin2, LocationConsumer2,
    * @param styleDelegate
    */
   override fun onStyleChanged(styleDelegate: StyleInterface) {
-    // no-ops
+    locationPuckManager?.updateStyle(styleDelegate)
   }
 
   override lateinit var internalSettings: LocationComponentSettings