+++ plugin-locationcomponent/src/main/java/com/mapbox/maps/plugin/locationcomponent/LocationLayerRenderer.kt
@@ -38,4 +38,6 @@ internal interface LocationLayerRenderer {
   )
 
   fun clearBitmaps()
+
+  fun updateStyle(style: StyleInterface)
 }